Check if option exists, before calling `delete_option`
Helps with narrowing down the number of SQL queries triggered on every admin screen.

fixes #30823

+++ plugins/woocommerce/includes/admin/class-wc-admin-notices.php
@@ -141,7 +141,9 @@ class WC_Admin_Notices {
 	 */
 	public static function remove_notice( $name, $force_save = false ) {
 		self::$notices = array_diff( self::get_notices(), array( $name ) );
-		delete_option( 'woocommerce_admin_notice_' . $name );
+		if ( false !== get_option( 'woocommerce_admin_notice_' . $name ) ) {
+			delete_option( 'woocommerce_admin_notice_' . $name );
+		}
 
 		if ( $force_save ) {
 			// Adding early save to prevent more race conditions with notices.
